# swagger.yaml
swagger: "2.0"
info:
  description: "My hockey teams application"
  version: "1.0.0"
  title: "Hockey teams"
  termsOfService: "https://google.com"
  contact:
    email: "p_perestoronin@mail.ru"
host: "localhost:3000"
basePath: "/api"
...
schemes:
- "https"
- "http"
paths:
  /player:
    post:
      tags:
      - "player"
      summary: "Add a new player to the global base of players"
      description: ""
      security:
        - cookieAuth: []
      operationId: "addPlayer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Player object that needs to be added to global base (only for admins)"
        required: true
        schema:
          $ref: "#/definitions/Player"
      responses:
        "200":
          description: "Ok"
        "403":
          description: "No enough rights"
        "405":
          description: "Invalid input"

    get:
      tags:
      - "player"
      summary: "Get all players from the global base of players"
      description: ""
      operationId: "getPlayers"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Player"
  /player/{playerId}:
    patch:
      tags:
      - "player"
      summary: "Updates a player in the global base of players (only for admins)"
      description: ""
      security:
        - cookieAuth: []
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "playerId"
        in: "path"
        description: "ID of player that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
      - name: "body"
        in: "body"
        description: "Updated info of the player"
        required: true
        schema:
          $ref: "#/definitions/PlayerUpdInfo"
      responses:
        "200":
          description: "Player updated"
        "403":
          description: "No enough rights"
        "405":
          description: "Invalid input"
    get:
      tags:
      - "player"
      summary: "Find player by ID"
      description: "Returns a single player"
      operationId: "getPlayerById"
      produces:
      - "application/json"
      parameters:
      - name: "playerId"
        in: "path"
        description: "ID of player to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Player"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Player not found"

    delete:
      tags:
      - "player"
      summary: "Deletes a player"
      description: ""
      security:
        - cookieAuth: []
      operationId: "deletePlayer"
      produces:
      - "application/json"
      parameters:
      - name: "playerId"
        in: "path"
        description: "Player id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Player not found"
...
  /user:
    post:
      tags:
      - "user"
      summary: "Create user"
      description: ""
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/UserWithPass"
      responses:
        default:
          description: "successful operation"

    put:
      tags:
      - "user"
      summary: "Update user"
      description: ""
      operationId: "updateUser"
      security:
        - cookieAuth: []
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/UserWithPass"
      responses:
        "200":
          description: "Ok"
        "403":
          description: "No enough rights"
        "405":
          description: "Invalid input"
...
definitions:
...
  UserWithPass:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      login:
        type: "string"
      plevel:
        type: "integer"
        format: "int64"
      password:
        type: "string"
...
  Player:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      fname:
        type: "string"
      lname:
        type: "string"
      cntry:
        type: "string"
      dob:
        type: "string"
        format: "date-time"
  PlayerUpdInfo:
    type: "object"
    properties:
      fname:
        type: "string"
      lname:
        type: "string"
      cntry:
        type: "string"
securityDefinitions:
  cookieAuth:
    type: "apiKey"
    in: "header"
    name: "cookie"
...